backend/requirements.txt
Flask>=2.0
Flask-Cors
python-dotenv
marshmallow
pymongo
bcrypt
PyJWT
werkzeug==2.0.2
httpx==0.27.0
redis>=4.5.0 